refactor: consolidate track metadata structures into SimStructures namespace
